teht 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -46,12 +46,23 @@ blogsRouter.post('/', async (request, response, next) => {
 
 }) 
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const id = request.params.id
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const id = request.params.id
 
-  await Blog.findByIdAndRemove(id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(id)
   
-  response.status(204).end()
+    if (user.id.toString() !== blog.user.toString()) {
+      return response.status(401).json('invalid credentials')
+    }
+
+    await Blog.findByIdAndRemove(id)
+    response.status(204).end()
+  } catch (error) {
+    next(error)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
